requirements.txt
langchain-google-genai
langchain-core
langgraph
streamlit
python-dotenv
pypdf
pillow
kagglehub
pydantic